[fix] controller에 있는 scan, apply 로직을 service 패키지로 구조 변경
